requirements.txt
streamlit
pandas
numpy
matplotlib
faiss-cpu
torch
sentence-transformers
reportlab
requests
plotly
pytest
huggingface_hub
